ml/api/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
blis==0.7.11
catalogue==2.0.10
catboost==1.2.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
cmdstanpy==1.2.2
comm==0.2.2
confection==0.1.5
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
DAWG-Python==0.7.2
DAWG2-Python==0.8.0
debugpy==1.8.1
decorator==5.1.1
docopt==0.6.2
docopt-ng==0.9.0
ecdsa==0.18.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
holidays==0.45
huggingface-hub==0.23.3
idna==3.6
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.22.2
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
keras==3.3.3
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
llvmlite==0.42.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.3.0
mplcyberpunk==0.7.1
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
openpyxl==3.1.3
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
packaging==24.0
pandas==2.2.1
parso==0.8.3
passlib==1.7.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.22.0
preshed==3.0.9
prompt-toolkit==3.0.43
prophet==1.1.5
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
pymystem3==0.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rsa==4.9
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0-py3-none-any.whl#sha256=4b1c291c9b8991b708dadbf1b59f0fdaf7e70ca25e94df7fd3d523c4d5049754
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data==0.6.3
stanio==0.5.0
starlette==0.36.3
statsmodels==0.14.1
stumpy==1.12.0
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
thefuzz==0.22.1
thinc==8.2.4
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.41.2
tsfel==0.1.7
tsfresh==0.20.2
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
XlsxWriter==3.2.0